+++ lib/access_token_agent/missing_access_token.rb
@@ -1,0 +1,7 @@
+module AccessTokenAgent
+  class MissingAccessToken < Error
+    def initialize
+      super('The access token response did not contain an access token.')
+    end
+  end
+end