+++ lib/access_token_agent/token.rb
@@ -1,4 +1,5 @@
-require 'access_token_agent/invalid_token_type_error'
+require 'access_token_agent/missing_access_token'
+require 'access_token_agent/unsupported_token_type_error'
 
 module AccessTokenAgent
   class Token
@@ -7,15 +8,25 @@ module AccessTokenAgent
     EXPIRY_MARGIN = 60 # seconds
 
     def initialize(auth_response)
-      unless auth_response['token_type'] == 'bearer'
-        raise InvalidTokenTypeError, auth_response['token_type']
-      end
+      validate_response(auth_response)
+
       @value = auth_response['access_token']
       @expires_at = Time.now + auth_response['expires_in']
     end
 
     def valid?
       @expires_at - EXPIRY_MARGIN > Time.now
+    end
+
+    private
+
+    def validate_response(auth_response)
+      unless auth_response['token_type'] == 'bearer'
+        raise UnsupportedTokenTypeError, auth_response['token_type']
+      end
+
+      token = auth_response['access_token']
+      raise MissingAccessToken if token.nil? || token.empty?
     end
   end
 end